Fix bugs in User class, close #12 and close #13
Fixed bug related to #12. We can now load a user who has flashcards,
but who has no answer history for those cards. Also fixed bug related
to #13. The program can be invoked from either an IDE or on the command
line, and the location of the load and save files will not change
depending on where the program is run.