Remove call to encrypt password; we are still working on #17
We are not yet ready to encrypt the user's password and store it in
that format. There was a stray call to encrypt the user's password left
in the code from the earlier efforts on #17. I removed it so the program
will work as before until we complete this issue.